refactor: major architecture improvements and code cleanup
- Extract REST API routes from server.js into modular route modules:
  - backend/routes/admin.js - Admin dashboard API endpoints
  - backend/routes/leaderboard.js - Leaderboard API endpoints
  - backend/routes/analytics.js - Analytics tracking endpoints
  - backend/routes/geolocation.js - Geolocation API endpoint
  - backend/routes/health.js - Health check and metrics endpoints
  - backend/routes/index.js - Route registration entry point

- Add Zod validation schemas for socket events:
  - backend/utils/socketValidation.js - Comprehensive validation schemas
  - Integrated validation into wordHandler.js for submitWord, submitWordVote,
    and submitPeerValidationVote events

- Clean up legacy code:
  - Remove unused handlers_improved.js and socketHandlers.legacy.js
  - Remove legacy compatibility aliases from gameStateManager.js
  - Keep only actively used tournament management functions

- Security improvements:
  - CORS now fails closed in production (rejects wildcard)
  - Add proper error messages for validation failures

- server.js reduced from ~1340 lines to ~440 lines (67% reduction)
- Total code reduction: ~4,200 lines of dead code removed
